backend/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
Pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.0

# Media processing
ffmpeg-python>=0.2.0
pymediainfo>=6.1.0
piexif>=1.1.3

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Structured data processing
pandas>=2.1.0

# AI/ML - Embeddings
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
sentence-transformers>=2.2.0
ftfy>=6.1.1
regex>=2023.10.3

# Storage - Object Storage (S3)
boto3>=1.29.7
botocore>=1.32.7

# Database - MongoDB
pymongo>=4.6.0
motor>=3.3.2

# Database - Supabase (alternative)
supabase>=2.3.0

# Vector Database - Pinecone
pinecone-client>=3.0.0

# API Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Utilities
aiofiles>=23.2.1
tenacity>=8.2.3
tqdm>=4.66.1


# Document processing
PyPDF2==3.0.1
python-docx==1.2.0
requests
pillow-heif>=0.10.0
